Create Coursemapping University of Vienna
This searches items that I identified as courses due to their hyperlink being preceded by a code e.g. "280299 VO". Other items in the directory seemed to be links to other semesters, degrees, etc, never leading to a course description. Other items were furthermore not consistent in terms of html. This search may not find all results due to this.